data-stack/to_be_deleted/ibc2025-data-gen-acme-video-telemetry-synthetic/cdk/requirements.txt
aws-cdk-lib==2.150.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-msk-alpha==2.150.0a0